+++ Assets/Scripts/Entity/EntityHealth.cs
@@ -7,7 +7,7 @@ public class EntityHealth : MonoBehaviour
 {
 	[SerializeField] private Text text_Health;
 
-	public void UpdateHealthText(int amount)
+	public void UpdateHealthText(float amount)
 	{
 		text_Health.text = amount.ToString();
 	}
